+++ src/app/eth-common/module/dashboard/INetworkModuleOptions.tsx
@@ -1,0 +1,7 @@
+export interface INetworkModuleOptions {
+    networkName: string;
+    otherNetworks?: {
+        name: string;
+        url: string;
+    }[];
+}