+++ packages/eth-common/CHANGELOG.md
@@ -1,6 +1,7 @@
 # v2.1.0
 
 - Add toolbar product links from `@alethio/explorer-plugin-eth-extended`. See [config.sample.json](config.sample.json) for configuration changes.
+- Add module showing the current network and option to switch to other deployments or external URLs
 
 # v2.0.0